ps_madd rounded the product twice; 354 wrong values become 350
The gate I rewrote an hour ago said teaching report.py to split value
failures from FPSCR-state failures was the first task. That was wrong: the
split already exists as a second assembler build, documented in the
Makefile I had already read.

  make clean && make EXTRA_DEFSYM=-Wa,--defsym,IGNORE_FPSCR_STATE=1

That gives exactly 354, and only then does report.py's ranking mean
anything. The default build's table is dominated by the 676 state failures
and ranks frsp first, which is a state failure -- which is how I spent an
hour on frsp yesterday.

Real ranking of the 354: ps_nmadd 18, psq_st 16, ps_nmsub 16, frsp 15,
ps_cmpo1 14, ps_mul 14, fctiw 13, and 131 in the multiply-add family.

PS_MADD wrapped the product in its own (float) cast before adding, so the
intermediate was rounded to single twice. PowerPC multiply-add is fused;
the product must not be rounded. Every sibling -- PS_NMADD, PS_MSUB,
PS_NMSUB, and FMADDS -- already casts only the whole expression, so
ps_madd was the one member of its own family doing it wrong. The suite's
changelog names the case verbatim: "fmadds/ps_madd is not rounded twice".

Values 354 -> 350, ps_madd 16 -> 12.

The FULL suite stays at 1030 and that is not a contradiction: those four
tests still fail their FPSCR-state check, which is the deliberate
omission. The headline number cannot show this fix, which is the argument
for measuring against the values-only build.

The siblings already have the correct fused shape, so whatever is wrong
with them is something else. That is the next thing to isolate.

+++ src/Cafe/HW/Espresso/Interpreter/PPCInterpreterPS.cpp
@@ -73,8 +73,12 @@ void PPCInterpreter_PS_MADD(PPCInterpreter_t* hCPU, uint32 Opcode)
 	frA = (Opcode>>16)&0x1F;
 	frD = (Opcode>>21)&0x1F;
 
-	float s0 = (float)((float)(hCPU->fpr[frA].fp0 * roundTo25BitAccuracy(hCPU->fpr[frC].fp0)) + hCPU->fpr[frB].fp0);
-	float s1 = (float)((float)(hCPU->fpr[frA].fp1 * roundTo25BitAccuracy(hCPU->fpr[frC].fp1)) + hCPU->fpr[frB].fp1);
+	// ps_madd is FUSED: the product is not rounded before the add. The inner (float) cast here
+	// rounded it, making this the one member of the family that rounds twice -- every sibling
+	// (PS_NMADD, PS_MSUB, PS_NMSUB) already casts only the whole expression, as does FMADDS.
+	// ppc750cl.s's own changelog names this case: 'fmadds/ps_madd is not rounded twice'.
+	float s0 = (float)(hCPU->fpr[frA].fp0 * roundTo25BitAccuracy(hCPU->fpr[frC].fp0) + hCPU->fpr[frB].fp0);
+	float s1 = (float)(hCPU->fpr[frA].fp1 * roundTo25BitAccuracy(hCPU->fpr[frC].fp1) + hCPU->fpr[frB].fp1);
 
 	hCPU->fpr[frD].fp0 = flushDenormalToZero(s0);
 	hCPU->fpr[frD].fp1 = flushDenormalToZero(s1);
